+++ tests/Expect/Time/toBeFuture.php
@@ -1,0 +1,19 @@
+<?php
+
+use PHPUnit\Framework\ExpectationFailedException;
+
+test('pass', function () {
+    expect(now()->addHour())->toBeFuture();
+});
+
+test('fails', function () {
+    expect('2999-01-01')->toBeFuture();
+})->throws(ExpectationFailedException::class, 'Failed to assert that [2999-01-01 00:00:00] is in the future');
+
+test('pass negated', function () {
+    expect(now())->not->toBeFuture();
+});
+
+test('fails negated', function () {
+    expect(now()->addHour())->not->toBeFuture();
+})->throws(ExpectationFailedException::class);